refactor(webhook): move reply token handling and context management to utils - so that we can share them with other handlers
